ENH: Add CI for testing via GH actions
This commit adds a workflow file for Github actions to run the package tests
on pull requests and pushes to the master branch on a ubuntu-latest runner.

Additionally,the minimum Python version is increased to 3.10, to support
typing, and the tests are run for 3.10 and the latest stable Python version.